+++ lumenairy/propagators/subaperture.py
@@ -31,7 +31,7 @@ from typing import List, Tuple
 
 import numpy as np
 
-from ._array import array_namespace, is_jax_array
+from ..backend import array_namespace, is_jax_array
 
 
 @dataclass